+++ benchopt/cli/tests/test_cli.py
@@ -6,9 +6,7 @@ import click
 import pytest
 from joblib.memory import _FUNCTION_HASHES
 
-from benchopt.plotting import PLOT_KINDS
 from benchopt.utils.temp_benchmark import temp_benchmark
-
 
 from benchopt.tests.utils import CaptureCmdOutput
 from benchopt.tests.utils import patch_var_env
@@ -576,27 +574,55 @@ class TestPlotCmd:
             plot(f"{self.bench.benchmark_dir} -k invalid_kind --html "
                  f"--no-display".split(), 'benchopt', standalone_mode=False)
 
-    @pytest.mark.parametrize('kind', list(PLOT_KINDS.keys()) + ["custom_plot"])
-    def test_valid_call(self, kind):
+    @pytest.mark.parametrize(
+        ('kind', 'expected_n_files'),
+        [
+            ("custom_plot", 1),
+            ("objective_curve", 2),
+            ("boxplot", 4),
+            ("bar_chart", 1),
+            (None, 8)  # all kinds
+        ]
+    )
+    def test_valid_call_mpl(self, kind, expected_n_files):
 
         with CaptureCmdOutput() as out:
-            plot(f"{self.bench.benchmark_dir} -f {self.result_file} -k {kind} "
-                 "--no-display --no-html".split(),
-                 'benchopt', standalone_mode=False)
+            cmd = f"{self.bench.benchmark_dir} -f {self.result_file} "
+            cmd += "--no-display --no-html "
+            if kind is not None:
+                cmd += f"--kind {kind}"
+            plot(cmd.split(), 'benchopt', standalone_mode=False)
 
-        assert len(out.result_files) == 1
-        assert kind in out.result_files[0]
-        assert '.pdf' in out.result_files[0]
+        assert len(out.result_files) == expected_n_files
+        for file in out.result_files:
+            if kind is not None:
+                assert kind in file
+            assert '.pdf' in file
 
-    def test_valid_call_html(self):
+    @pytest.mark.parametrize(
+        'kind',
+        ["custom_plot", "objective_curve", "boxplot", "bar_chart", None]
+    )
+    def test_valid_call_html(self, kind):
 
-        with CaptureCmdOutput() as out:
-            plot(f"{self.bench.benchmark_dir} -f {self.result_file} "
-                 "--no-display --html".split(),
-                 'benchopt', standalone_mode=False)
+        with CaptureCmdOutput(delete_result_files=False) as out:
+            cmd = f"{self.bench.benchmark_dir} -f {self.result_file} "
+            cmd += "--no-display --html "
+            if kind is not None:
+                cmd += f"--kind {kind}"
+            plot(cmd.split(), 'benchopt', standalone_mode=False)
 
         assert len(out.result_files) == 2
         assert all('.html' in f for f in out.result_files)
+
+        html_content = Path(out.result_files[0]).read_text()
+        for k in [
+            "custom_plot", "objective_curve", "boxplot", "bar_chart"
+        ]:
+            if kind is None or k == kind:
+                assert f"<option value=\"{k}\"" in html_content
+            else:
+                assert f"<option value=\"{k}\"" not in html_content
 
     def test_complete_bench(self, bench_completion_cases):  # noqa: F811