+++ benchopt/constants.py
@@ -1,5 +1,0 @@
-PLOT_KINDS = {
-    # 'objective_curve': 'plot_objective_curve',
-    'bar_chart': 'plot_bar_chart',
-    'boxplot': 'plot_boxplot'
-}