+++ benchopt/plotting/__init__.py
@@ -1,4 +1,3 @@
-import itertools
 import pandas as pd
 import matplotlib.pyplot as plt
 
@@ -6,10 +5,6 @@ import matplotlib.pyplot as plt
 from ..utils.parquet import get_metadata
 from ..utils.parquet import update_metadata
 
-from ..constants import PLOT_KINDS
-from .helpers import get_plot_id
-from .plot_boxplot import plot_boxplot  # noqa: F401
-from .plot_bar_chart import plot_bar_chart  # noqa: F401
 from .generate_html import plot_benchmark_html
 from .generate_matplotlib import get_figures
 
@@ -77,17 +72,14 @@ def plot_benchmark(fname, benchmark, kinds=None, display=True, plotly=False,
     if kinds is not None and len(kinds) > 0:
         config["plots"] = kinds
 
+    valid_kinds = benchmark.get_plot_names()
+
     if "plots" not in config or config["plots"] is None:
-        config["plots"] = (
-            benchmark.get_default_plot_names() +
-            list(PLOT_KINDS.keys()) +
-            benchmark.get_custom_plot_names()
-        )
+        config["plots"] = valid_kinds
 
     if html:
         plot_benchmark_html(fname, benchmark, config, display)
 
-    # TODO: rewrite this when default plots are also custom plots
     else:
         # Load the results.
         if fname.suffix == '.parquet':
@@ -97,18 +89,8 @@ def plot_benchmark(fname, benchmark, kinds=None, display=True, plotly=False,
         if "data_name" in df.columns:
             df = df.rename(columns={"data_name": "dataset_name"})
 
-        obj_cols = [
-            k for k in df.columns
-            if k.startswith('objective_') and k != 'objective_name'
-        ]
-        datasets = df['dataset_name'].unique()
         output_dir = benchmark.get_output_folder()
 
-        valid_kinds = (
-            benchmark.get_default_plot_names() +
-            list(PLOT_KINDS.keys()) +
-            benchmark.get_custom_plot_names()
-        )
         for kind in config["plots"]:
             if kind not in valid_kinds:
                 raise ValueError(
@@ -123,40 +105,6 @@ def plot_benchmark(fname, benchmark, kinds=None, display=True, plotly=False,
         )
         figs.extend(kind_figs)
 
-        for data in datasets:
-            df_data = df[df['dataset_name'] == data]
-            objective_names = df['objective_name'].unique()
-            for objective_name in objective_names:
-                df_obj = df_data[df_data['objective_name'] == objective_name]
-
-                plot_id = get_plot_id(benchmark.name, df_obj)
-
-                for kind, obj_col in itertools.product(
-                        config["plots"], obj_cols
-                ):
-                    if kind not in PLOT_KINDS:
-                        continue
-                    # For now only plot bar chart and suboptimality for
-                    # objective_value for which we monitor convergence
-                    # XXX - find a better solution
-                    if obj_col != "objective_value" and (
-                            kind == "bar_chart" or "subopt" in kind):
-                        continue
-                    plot_func = globals()[PLOT_KINDS[kind]]
-                    try:
-                        fig = plot_func(df_obj, obj_col=obj_col, plotly=plotly)
-                    except TypeError:
-                        fig = plot_func(df_obj, obj_col=obj_col)
-                    save_name = output_dir / f"{plot_id}_{obj_col}_{kind}"
-                    if hasattr(fig, 'write_html'):
-                        save_name = save_name.with_suffix('.html')
-                        fig.write_html(str(save_name), include_mathjax='cdn')
-                    else:
-                        save_name = save_name.with_suffix('.pdf')
-                        plt.savefig(save_name)
-                    print(f'Save {kind} plot of {obj_col} for {data} and '
-                          f'{objective_name} as: {save_name}')
-                    figs.append(fig)
         if display:
             plt.show()
         return figs