+++ examples/plot_run_benchmark_python_R.py
@@ -9,8 +9,8 @@ from pathlib import Path
 import matplotlib.pyplot as plt
 from benchopt import run_benchmark
 from benchopt.benchmark import Benchmark
-from benchopt.plotting import plot_benchmark, PLOT_KINDS
-from benchopt.plotting.helpers import reset_solver_styles_idx
+from benchopt.plotting import plot_benchmark
+from benchopt.plotting.helpers import reset_solver_styles
 
 
 BENCHMARK_PATH = Path().resolve().parent / 'benchmarks' / 'benchmark_lasso'
@@ -32,10 +32,9 @@ _, save_file = run_benchmark(
     plot_result=False, show_progress=False
 )
 
+reset_solver_styles()
 
-kinds = list(PLOT_KINDS.keys())
-reset_solver_styles_idx()
 figs = plot_benchmark(
-    save_file, benchmark=Benchmark(BENCHMARK_PATH), kinds=kinds, html=False
+    save_file, benchmark=Benchmark(BENCHMARK_PATH), html=False
 )
 plt.show()